# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bill's Pro API",
        "description": "API documentation for Bill's Pro - Crypto and Bill Payment System",
        "contact": {
            "name": "API Support",
            "email": "support@billspro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/bill-payment/categories": {
            "get": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "Get bill payment categories",
                "description": "Get all available bill payment categories.",
                "operationId": "a9fe7cb09366e85fa10f04a66d624715",
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bill-payment/providers": {
            "get": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "Get providers by category",
                "description": "Get providers for a specific category.",
                "operationId": "8dbf34edefdf8b037e06ba2da1f91ca1",
                "parameters": [
                    {
                        "name": "categoryCode",
                        "in": "query",
                        "description": "Category code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "airtime"
                        }
                    },
                    {
                        "name": "countryCode",
                        "in": "query",
                        "description": "Country code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "NG"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Providers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Category code is required"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bill-payment/plans": {
            "get": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "Get plans by provider",
                "description": "Get available plans/bundles for a provider (Data and Cable TV only).",
                "operationId": "6a54dff46b1b0be4a4ece284bfd4352f",
                "parameters": [
                    {
                        "name": "providerId",
                        "in": "query",
                        "description": "Provider ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plans retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Provider ID is required"
                    },
                    "404": {
                        "description": "Provider not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bill-payment/validate-meter": {
            "post": {
                "tags": [
                    "Bill Payment"
                ],
                "summary": "Validate electricity meter",
                "description": "Validate electricity meter number before payment.",
                "operationId": "4628983512ab632a00d0fecb9094ff22",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "providerId",
                                    "meterNumber",
                                    "accountType"
                                ],
                                "properties": {
                                    "providerId": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "meterNumber": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "accountType": {
                                        "type": "string",
                                        "enum": [
                                            "prepaid",
                                            "postpaid"
                                        ],
                                        "example": "prepaid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Meter validated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid meter number"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}